makeup-rcmd-products/requirements.txt
streamlit
pandas
numpy
scikit-learn
joblib
python-dotenv
groq
shap
matplotlib
